Use RequireJS with or without config file to resolve module paths
* Load require.js using npm module dependency.
* Use the configuration property requireModuleSupport.modulePath as
baseUrl for require.config().
* If the configuration property requireModuleSupport.configFile is set,
execute this file to initialize the default requirejs context.
* Introduce the resolveModulePath method resolving the module path.
* Use requirejs.toUrl() to resolve the file-system path to the module
file. If the module path is plugin expression, cut the plugin name and
the exclamation mark and handle the rest of the expression as another
module path.
* If the configuration property requireModuleSupport.pluginExtensions is
set, ensure, that the module path after cutting the plugin name has the
file extension as specified in the configuration object.
* Document the new configuration properties in the README.